@@ -59,60 +67,61 @@
     "leaflet.markercluster": "^1.4.1",
     "leaflet.zoominfo": "git+https://github.com/SemanticComputing/Leaflet.zoominfo.git",
     "lodash": "^4.17.21",
-    "mapbox-gl-leaflet": "0.0.15",
+    "mapbox-gl-leaflet": "^0.0.15",
     "moment": "^2.29.2",
     "moment-range": "^4.0.2",
     "process": "^0.11.10",
+    "prop-types": "^15.8.1",
     "react": "^17.0.2",
     "react-csv": "^2.0.3",
     "react-dom": "^17.0.2",
-    "react-intl-universal": "^2.4.8",
-    "react-map-gl": "^6.1.0",
-    "react-redux": "^7.2.0",
-    "react-redux-toastr": "^7.6.11",
+    "react-intl-universal": "^2.9.1",
+    "react-map-gl": "^6.1.21",
+    "react-redux": "^7.2.9",
+    "react-redux-toastr": "^7.6.13",
     "react-router-dom": "^5.3.0",
     "react-sortable-tree-theme-file-explorer": "git+https://github.com/SemanticComputing/react-sortable-tree-theme-file-explorer.git#88cee6950b0656452987c5bd7e76780c62b1e913",
     "react-virtualized": "^9.22.3",
     "react-wordcloud": "^1.2.7",
-    "redux": "^4.1.2",
+    "redux": "^4.2.1",
     "redux-observable": "^2.0.0",
-    "reselect": "^4.0.0",
+    "reselect": "^4.1.8",
     "rxjs": "^7.2.0",
     "simple-react-lightbox": "^3.6.9-0",
     "simple-statistics": "^7.7.0",
-    "swagger-ui-express": "^4.1.6",
+    "swagger-ui-express": "^4.6.3",
     "tippy.js": "^6.3.7"
   },
   "devDependencies": {
-    "@babel/cli": "^7.16.0",
-    "@babel/core": "^7.15.5",
-    "@babel/eslint-parser": "^7.15.4",
-    "@babel/node": "^7.15.4",
-    "@babel/plugin-transform-runtime": "^7.15.0",
-    "@babel/preset-env": "^7.15.4",
-    "@babel/preset-react": "^7.14.5",
-    "@fortawesome/fontawesome-svg-core": "^1.2.34",
-    "@fortawesome/free-solid-svg-icons": "^5.15.2",
-    "babel-loader": "^8.2.2",
+    "@babel/cli": "^7.23.4",
+    "@babel/core": "^7.23.6",
+    "@babel/eslint-parser": "^7.23.3",
+    "@babel/node": "^7.22.19",
+    "@babel/plugin-transform-runtime": "^7.23.6",
+    "@babel/preset-env": "^7.23.6",
+    "@babel/preset-react": "^7.23.3",
+    "@fortawesome/fontawesome-svg-core": "^6.5.1",
+    "@fortawesome/free-solid-svg-icons": "^6.5.1",
+    "babel-loader": "^8.3.0",
     "clean-webpack-plugin": "^3.0.0",
     "compression-webpack-plugin": "^7.1.2",
-    "concurrently": "^6.3.0",
+    "concurrently": "^6.5.1",
     "copy-webpack-plugin": "^7.0.0",
     "cross-env": "^7.0.0",
-    "css-loader": "^5.0.1",
+    "css-loader": "^5.2.7",
     "dotenv": "^10.0.0",
     "file-loader": "^6.2.0",
     "googleapis": "^71.0.0",
-    "html-webpack-plugin": "^5.3.1",
-    "sass": "^1.43.4",
-    "sass-loader": "^10.2.0",
-    "sitemap": "^6.3.5",
+    "html-webpack-plugin": "^5.5.4",
+    "sass": "^1.69.5",
+    "sass-loader": "^10.5.0",
+    "sitemap": "^6.4.0",
     "standard": "^16.0.4",
     "style-loader": "^2.0.0",
-    "webpack": "^5.64.4",
-    "webpack-cli": "^4.8.0",
-    "webpack-dev-server": "^4.6.0",
-    "webpack-merge": "^5.7.3"
+    "webpack": "^5.89.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-server": "^4.15.1",
+    "webpack-merge": "^5.10.0"
   },
   "standard": {
     "parser": "@babel/eslint-parser"
